refactor(test): MemberRestAssured로 통신하는 함수분리
